Test code separation
