--- AutoDoApp/parser/git_handler/tests.py
@@ -1,13 +1,0 @@
-# tests.py
-# unit test code for git_handler package
-# Author: Junho Kim
-
-import unittest
-
-
-class TestingUtils(unittest.TestCase):
-
-    test_url = "https://gtihub.com/JunoJunho/MathProject"
-
-    def setUp(self):
-        pass
